requirements.txt
azure-identity
azure-storage-blob
requests
